requirements.txt
openpyxl==3.1.5
pycryptodome==3.21.0
requests==2.32.3
